refactor: Small auth changes
I found it was better to change the auth status to retry the
account creation check. By doing it this way the clients doesn't
need to handle clearing timeouts and such.
